requirements.txt
amqp==2.4.1
anyjson==0.3.3
backcall==0.1.0
billiard==3.5.0.5
bleach==3.1.0
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
colorlover==0.3.0
cufflinks==0.14.6
decorator==4.3.2
defusedxml==0.5.0
Deprecated==1.2.4
dj-database-url==0.5.0
Django==1.11.8
django-bootstrap3==11.0.0
django-celery==3.2.2
django-celery-beat==1.1.1
django-celerybeat-status==0.0.10
django-crispy-forms==1.7.2
django-mysql==2.4.1
django-nose==1.4.6
entrypoints==0.3
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
kombu==4.3.0
MarkupSafe==1.1.0
mistune==0.8.4
mysqlclient==1.4.1
nbconvert==5.4.0
nbformat==4.4.0
nose==1.3.7
notebook==5.7.4
numpy==1.16.0
pandas==0.24.0
pandocfilters==1.4.2
parso==0.3.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pkg-resources==0.0.0
plotly==3.5.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
psycopg2==2.7.7
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
PyGithub==1.43.4
Pygments==2.3.1
PyJWT==1.7.1
pymongo==3.7.2
python-dateutil==2.7.5
python-nvd3==0.15.0
python-slugify==2.0.1
pytz==2018.9
pywinpty==0.5.5
pyzmq==17.1.2
requests==2.21.0
retrying==1.3.3
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
Unidecode==1.0.23
urllib3==1.24.1
vine==1.2.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wrapt==1.11.1
